Keep ::Plugin::File from wrongly using --literal
Unless --file is present, --literal needs to be passed through to ack.
Rather than consume it and add it back, I decided not to consume it at
all until we knew we had --file. This changes the point at which it is
removed from the command line, hence the change in t/plugin_file.t.