Clean up dependencies a little.
Re-implement __die() and __warn(), rather than aliasing App::Ack::die()
and App::Ack::warn(). This is the start of work on interfacing with the
single-file version of ack.

Load App::Ack::Filter modules with Module::Pluggable. This may be the
wrong way to go if the single-file solution involves putting ack _and_
ackxp in a single file.