Most syntax filters to while ( <$fh> ) {...} logic
The exception is ::Syntax::Data, where the $self->{want} logic moved to
the instantiator, since the only syntax type in the file _is_ data. The
previous logic was a bit strange, and involved a single-iteration loop
and 'redo'.

Also went to state machines for _cc_like and _single_line_comments.