Move PUSHED and FILL to ::Syntax.
These call __init() and __classify() respectively, which the subclasses
implement.

Each level of the inheritance hierarchy has its own attributes hash.
This is retrieved by __my_attr(), which uses caller() to associate the
hash with the caller's name space.

Thought: this way of doing things suggests basing the object on a code
reference, which would be an anonymous version of __my_attr(), closing
over the "master hash" that contains all the others.