Add --syntax-empty-code-is-comment
Normally empty lines are considered part of the syntax in which they are
embedded. If this Boolean option is asserted, empty code lines are
considered comments.

Note that the implementation is at the level of
App::Ackx::Preflight::Syntax -- the various __classify() methods still
return 'code' for a blank line embedded in code.