Add syntax support for the Crystal language.
This is incomplete, because the difference between documentation and
comments depends on what comes after the block of '#' lines. To tell the
difference I would have to save $. and the file position, read forward
until I found the end, determine what I had, and then restore the file
position and $.. And I don't feel like doing that yet.