Add utility __set_sub_name().
This is a convenience that sets the subroutine name of a code reference.
This is done on a best-effort baseis so as not to actually introduce a
dependency. It uses Sub::Util or Sub::Name as available. If neither is
available it does nothing, but returns the code reference to comply with
the interface contract.